+++ app/src/main/java/woowacourse/movie/view/reservation/ReservationAdapter.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.view.reservation
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
