+++ app/src/main/java/woowacourse/movie/view/reservation/ReservationViewHolder.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.view.reservation
 
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.movie.databinding.ItemReservationBinding
