+++ app/src/main/res/layout/fragment_reservation_history.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="adapter"
-            type="woowacourse.movie.ReservationAdapter" />
+            type="woowacourse.movie.view.reservation.ReservationAdapter" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
